+++ src/components/ui-system/content/WavingHand.tsx
